[PATCH] Missing end tags in kernel-locking kerneldoc
From: Rusty Russell <rusty@rustcorp.com.au>

From: Adam Kropelin <akropel1@rochester.rr.com>

The new-and-improved kernel-locking kerneldoc seems to be missing some end
tags which causes 'make foodocs' to die.  I'm not sure if it's because of
my not-bleeding-edge docbook utils or if it's a genuine error.  Since most
ending tags are present I tend to think it's an error.