[PATCH] Fix Documentation/SubmittingPatches to use -p
From: "Martin J. Bligh" <mbligh@aracnet.com>

Patches are a damned sight easier to read if people use the '-p'
option to diff ... this generates output that looks like this:

 "@@ -323,6 +323,7 @@ void put_dirty_page(struct task_struct *"

for each block. This patch simply adds that to the documentation
file, in the hope of steering new users in the right direction.