[PATCH] Arrange for EFI-related code to be compiled away
From: David Mosberger <davidm@napali.hpl.hp.com>

There is some EFI-related code which is present in the ia64 build but is not
needed: variable efi_enabled is always zero.

The patch fiddles with the efi_enabled definition to arrange for
`efi_enabled' to be constant zero or constant one in those situations where
this can be guaranteed.