[PATCH] NFS/RPC modprobe -r sunrpc causes an oops
From: Steve Dickson <SteveD@redhat.com>

Here is a patch for the 2.6.1 kernel that fixes an oops that occurs when
the sunrpc module is unloaded.

The problem was the RPC cache_register() call was not saving entry pointers
to the procfs entries it was creating.  So when it came time to dismantle
the entires, a BUG_ON() was tripped in remove_proc_entry() since the tree
was not broken down completely.

(acked by neilb)