[PATCH] Fine tune the time conversion to eliminate conversion errors.
From: George Anzinger <george@mvista.com>

The time conversion code is erroring on the side of a bit too small.  The
attached patch forces any error to be on the high side.  The current code will
convert 1 nanosecond to zero jiffies (standard says that should be 1).  It also
is around 1 nanosecond late on each roll to the next jiffie.

I have done some error checks with this patch applied and get the following
errors in PPB ( Parts Per Billion):

HZ     nano sec conversion     microsecond conversion
1000    315                      45
1024    227                      40
100     28                       317

In all cases the error is on the high side, which means that the final shift
will, most likely, eliminate the error bits.